+++ src/editor/mod.rs
@@ -2,5 +2,6 @@ pub mod blocks;
 pub mod buffer;
 pub mod cursor;
 pub mod editor;
+pub mod highlight;
 
 pub use editor::Editor;